Fix - mirrmaid.conf requires too much editing to change source
I had it really close to being properly abstracted, but my mistake was
using a single prefix to be used for both source and target.  That is
the primary difference between mirrors.  Since I want my mirrors to
remain at a fixed location and have the ability to use a variety of
upstream mirrors it really is necessary to have a source prefix (that
varies by the upstream mirror) and a target prefix (simply for good
design).  This revision accomplishes that.

It also provides TDS as another mirror source.