Fix - cron may not run mirrmaid on fresh build
This continues in the vain of 86edc1 to ensure that the pre-install
scripts in the mirrmaid package will be able to successfully recognize
that a mirrmaid user and group already exist ... in LDAP.  It goes
further by notifying crond of the new account additions.

The last build of brian (f16) did not create a local mirrmaid user
account, but crond refused to run the job as the mirrmaid user was
considered invalid.  This rev is untested, but theoretically may help.